refactor(article): divide repository
