+++ src/APIs/articles/dtos/article-response.dto.ts
@@ -1,15 +1,15 @@
 import { ApiProperty, OmitType } from '@nestjs/swagger';
 
 import { UserPrimaryResponseDto } from 'src/APIs/users/dtos/user-response.dto';
-import { Posts } from '../entities/article.entity';
+import { Article } from '../entities/article.entity';
 
-export class PostResponseDto extends OmitType(Posts, ['user']) {
+export class ArticleDetailResponse extends OmitType(Article, [
+  'comments',
+  'user',
+  'stickerBlocks',
+  'reports',
+  'notifications',
+]) {
   @ApiProperty({ description: '작성자의 정보', type: UserPrimaryResponseDto })
   user: UserPrimaryResponseDto;
 }
-
-export class PostOnlyResponseDto extends OmitType(Posts, [
-  'user',
-  'postBackground',
-  'postCategory',
-]) {}
