+++ src/APIs/articles/dtos/article.dto.ts
@@ -1,4 +1,12 @@
 import { OmitType } from '@nestjs/swagger';
 import { Article } from '../entities/article.entity';
 
-export class ArticleDto extends OmitType(Article, ['comments', '']) {}
+export class ArticleDto extends OmitType(Article, [
+  'comments',
+  'user',
+  'stickerBlocks',
+  'reports',
+  'notifications',
+  'articleCategory',
+  'articleBackground',
+]) {}
