+++ src/APIs/articles/olds/fetch-posts.dto.ts
@@ -1,7 +1,7 @@
 import { IsEnum, IsOptional, IsString } from 'class-validator';
 import { PageRequest } from '../../../utils/pages/page-request';
 import { ApiProperty } from '@nestjs/swagger';
-import { PostsOrderOptionWrap } from 'src/common/enums/posts-order-option';
+import { PostsOrderOptionWrap } from 'src/common/enums/article-order-option';
 import { PostsFilterOptionWrap } from 'src/common/enums/posts-filter-option';
 
 export class FetchPostsDto extends PageRequest {
