+++ src/common/enums/article-filter-option.ts
@@ -1,10 +1,10 @@
-export enum PostsFilterOption {
+export enum ArticleFilterOption {
   TITLE = 'p.title',
   CONTENT = 'p.content',
   USER = 'user.username',
 }
 // client에게 key값을 받기 위한 wrapping enum
-export enum PostsFilterOptionWrap {
+export enum ArticleFilterOptionWrap {
   TITLE = 'TITLE',
   CONTENT = 'CONTENT',
   USER = 'USER',
