+++ src/common/enums/article-order-option.ts
@@ -1,4 +1,4 @@
-export enum PostsOrderOption {
+export enum ArticleOrderOption {
   LIKE = 'like_count',
   VIEW = 'view_count',
   COMMENT = 'comment_count',
@@ -6,7 +6,7 @@ export enum PostsOrderOption {
 }
 
 // client에게 key값을 받기 위한 wrapping enum
-export enum PostsOrderOptionWrap {
+export enum ArticleOrderOptionWrap {
   LIKE = 'LIKE',
   VIEW = 'VIEW',
   COMMENT = 'COMMENT',
